Update validation messages as per wireframe
From: https://eaflood.atlassian.net/browse/RUBY-923

Updates validation messages in registration journey to be in-line with the wireframe content
