+++ app/forms/waste_carriers_engine/base_form.rb
@@ -4,6 +4,7 @@ module WasteCarriersEngine
   class BaseForm
     include ActiveModel::Model
     include CanStripWhitespace
+    include CanSetCustomErrorMessage
 
     extend ActiveModel::Callbacks
