+++ app/forms/waste_carriers_engine/business_type_form.rb
@@ -4,6 +4,9 @@ module WasteCarriersEngine
   class BusinessTypeForm < BaseForm
     delegate :business_type, to: :transient_registration
 
-    validates :business_type, "defra_ruby/validators/business_type": { allow_overseas: true }
+    validates :business_type, "defra_ruby/validators/business_type": {
+      allow_overseas: true,
+      messages: custom_error_messages(:business_type, :inclusion)
+    }
   end
 end
