+++ config/locales/forms/business_type_forms/en.yml
@@ -19,6 +19,8 @@ en:
       models:
         waste_carriers_engine/business_type_form:
           attributes:
+            business_type:
+              inclusion: "Select type of business or organisation"
             reg_identifier:
               invalid_format: "The registration ID is not in a valid format"
               no_registration: "There is no registration matching this ID"
