+++ config/locales/forms/cbd_type_forms/en.yml
@@ -22,7 +22,7 @@ en:
         waste_carriers_engine/cbd_type_form:
           attributes:
             registration_type:
-              inclusion: "You must answer this question"
+              inclusion: "Select who carries the waste"
             reg_identifier:
               invalid_format: "The registration ID is not in a valid format"
               no_registration: "There is no registration matching this ID"
