+++ config/locales/forms/construction_demolition_forms/en.yml
@@ -18,7 +18,7 @@ en:
         waste_carriers_engine/construction_demolition_form:
           attributes:
             construction_waste:
-              inclusion: "You must answer this question"
+              inclusion: "Select yes or no"
             reg_identifier:
               invalid_format: "The registration ID is not in a valid format"
               no_registration: "There is no registration matching this ID"
