+++ config/locales/forms/declare_convictions_forms/en.yml
@@ -45,7 +45,7 @@ en:
         waste_carriers_engine/declare_convictions_form:
           attributes:
             declared_convictions:
-              inclusion: "You must answer this question"
+              inclusion: "Select yes or no"
             reg_identifier:
               invalid_format: "The registration ID is not in a valid format"
               no_registration: "There is no registration matching this ID"
