+++ config/locales/forms/service_provided_forms/en.yml
@@ -15,7 +15,7 @@ en:
         waste_carriers_engine/service_provided_form:
           attributes:
             is_main_service:
-              inclusion: "You must answer this question"
+              inclusion: "Select who creates the waste"
             reg_identifier:
               invalid_format: "The registration ID is not in a valid format"
               no_registration: "There is no registration matching this ID"
